refactor: Migrate orbit-policy tests to per-module tests/ layout [ORB-00229]
